Implement local filtering for Friends and Groups using StateNotifier (#82)
## Pull Request: Implement Local Filtering for Friends and Groups

### Overview
This Pull Request introduces significant improvements to the Friends and
Groups pages by implementing local filtering capabilities. This ensures
that the application remains responsive and reduces the need for
repeated API calls. The changes involve the integration of
`StateNotifier` to manage the lists' state based on user input.

### Key Changes
- **StateNotifiers Added**: `FriendsStateNotifier` and
`GroupsStateNotifier` have been introduced to handle the state
management of friends and groups lists.
- **Data Loading Optimization**: The initial data loading logic has been
centralized in the `AuthenticationPage`. It initializes the
StateNotifiers with the fetched data, ensuring the data is only loaded
once during the app lifecycle.
- **UI Updates**: Both the `FriendsPage` and `GroupsPage` have been
updated to use the newly created StateNotifiers for filtering. This
update includes modifications to the search bars to trigger state
updates on text changes.

### Benefits
- **Performance**: Reduces the load on the server by minimizing API
calls and handles data more efficiently on the client side.
- **User Experience**: Enhances the responsiveness of the application by
immediately updating the list based on search criteria without the need
for fetching new data.
- **Maintainability**: By using `StateNotifier`, the state management is
more structured and easier to maintain.

### Testing
Please ensure to test the following:
- The friends and groups lists update correctly based on the search
input.
- No additional API calls are made during filtering.
- The application does not reload data unnecessarily on re-navigations
to the same page.

### Video



https://github.com/ISIS3510-202410-Team-13/Flutter/assets/78111224/f2b92da5-f3dd-4340-b58c-4f48d7a90b56



### Additional Notes
- Please review the error handling and suggest any potential
improvements.
- Feedback on the UI changes and any additional features that could be
included is welcome.

---

Looking forward to your reviews and suggestions!