Initial Implementation of local storage strategy for events, friends and groups (Just Saving for now)  (#88)
Local storage strategy using Hive for saving data for events, friends,
and groups. Below are the detailed changes.

1. The models—Event, Friend, and Group—are now integrated with Hive,
allowing the app to store data locally. This enhancement is pivotal as
it provides offline capabilities and improves data retrieval
performance. Each model is annotated with @HiveType and fields are
marked with @HiveField to ensure correct serialization and
deserialization.
2. Changes in pages like AuthenticationPage reflect the integration with
Hive, using updated models to store data locally. State providers are
also adjusted to manage state using Hive,
3. Temporal GUI added to see the data saved in local storage related to
Friends, Groups and Events. It is accesible from the side-out menu in
the "About" section.

Here its a video for showing this:



https://github.com/ISIS3510-202410-Team-13/Flutter/assets/69651671/d8aaf0d9-b7d7-47a1-949d-3bc2b91f6e42

Closes #69

Co-authored-by: Sebastian Urrea <68788933+jsurrea@users.noreply.github.com>