Create personal events (#97)
I created new providers to manage the state of events. I used those
providers to create a very basic new event within the create-class page.
I also updated the UI to allow the rendering of any event. This handles:

- Any day of the week (Mon-Sun)
- Any hour of the day (00:00 - 23:59)
- Scroll to see any element
- Event name is displayed dynamically (I calculate appropriate number of
lines and use overflow with ellipsis)
- Data is fetched from the database


https://github.com/ISIS3510-202410-Team-13/Flutter/assets/68788933/08345b2b-740d-47c7-857e-af99244f2fe1

There are many things to consider in following implementations:
- How to handle stacking elements over the same hours
- Saving new events to the database
- **Differentiating recurrent vs one-time events** (this might involve
changes in the database)
- Include duration, color, labels, title, etc when creating the event

I will take care of more of these things in the following weeks. For
now, this version works good enough while I fix some issues related to
technical debt which are a must at this point!