Refactor layers as Clean Architecture  (#99)
# Overview
This PR implements a refactoring of most files to follow a consistent
structure.
Additionally, it solves a few bugs and leaves the ground ready for next
steps in eventual connectivity strategies. This is a brief overview of
the changes:

## Others
- Updated README with new libs structure, instructions on how to set up
Firebase and the command to use `build_runner` in watch mode.
- Updated gitignore to exclude files with API Keys for Firebase.
- Updated Android and iOS files with Firebase Config and authorization.
- Fixed Multidex errors on Android (caused when methods use more than a
fixed amount of memory in the JVM).
- Updated linter configs and removed unnecessary content in several
files generated when creating the project.

# Future Work
I left many improvements that can be implemented through the code marked
as **TODO** (easily searchable by the IDE). Some of the most important
things are:
- There are many components in `NewEventView` which can be refactored
into the `widgets` dir. This will enable us to reuse things such as
TextInputs, ColorPickers, DropdownSelectors, etc. in other views of the
application with ease The file is still very big and additional effort
is needed for it to become more manageable.
- Services related to Notifications and Local Auth can be improved with
design patterns such as Singletons. Currently, they're very diverse in
their implementation. The definition of a "common" API for them could be
implemented.
- Widgets related to Auth, Calendar, NewEvent and the Drawer (sidebar)
can be further refactored to use Constants. This is an extensive labor,
which I hope we can progress on through this sprint and the next one.