Local Storage for friends and groups, and text messages for status of the lists (#104)
# Implementation of Local Storage Strategy for Friends and Groups and
Connectivity Management of those.

### Now
Here you can see a video of how it works now:


https://github.com/ISIS3510-202410-Team-13/Flutter/assets/78111224/b028f3e7-27fb-42d0-8bf3-1dfbcdb96bf4

Now the Friends and Groups are storaged in Local Storage and shown if
there is not internet connection. In case there is not internet
connection and not local information is found, then it shows messages
like: "No friends to display at the moment. Please check your internet
connection to view your friends.", if there is an internet connection
but not local information is found it shows message like ""You currently
have no friends to display. Add one to see them here.""