Fix error when rendering profile icons in groups card when there are less than 3 members (#109)
When there are less than 3 members, the white background does not render
properly, causing Stack to throw an exception for undefined size
behavior. For this, I chose to leave the white background as-is, but
only show the text of remainder members when there is at least one. This
causes the row to have empty whitespace, as you can see in the image. We
can address this better later. For now, it works with any number of
members in the group.

I also changed the iterator to traverse the array up to the point where
there are still values. Previously, it was fixed to at least 3, causing
errors.

<img width="401" alt="image"
src="https://github.com/ISIS3510-202410-Team-13/Flutter/assets/68788933/811edb64-f5bc-41d0-8901-bbd14d416f31">