Chat Features (send and receive messages) (#153)
This PR introduces new features to send and receive chat messages.



https://github.com/ISIS3510-202410-Team-13/Flutter/assets/68788933/12e4d362-7303-45ee-8b72-4b1ad748584b

### 1. Firestore Schema for Chat Messages

In Firebase Firestore, a collection named "Chats" is used to store chat
messages between users. Each chat conversation is represented by a
document with a unique identifier, which is generated by sorting the
user IDs of the participants and concatenating them with an underscore.
For example, a chat between user A and user B would have a document ID
of "A_B".

Within each chat document, there is a subcollection named "messages"
that stores individual chat messages. Each message is represented by a
document containing the following fields:

- `message`: The text content of the message.
- `name`: The display name of the sender.
- `profilePicture`: The URL of the sender's profile picture.
- `senderId`: The unique identifier of the sender.
- `timestamp`: The timestamp indicating when the message was sent.

### 2. Chat Service Implementation

The `ChatService` class is responsible for interacting with Firebase
Firestore to send and receive chat messages. It provides two main
functionalities:

- **Sending Messages**: When a user sends a message, the `sendMessage`
method is called. This method constructs a `MessageModel` object with
the necessary message data and adds it to the corresponding chat
document's "messages" subcollection in Firestore.

- **Receiving Messages**: To retrieve chat messages for a specific
conversation, the `getMessages` method is used. This method listens to
changes in the Firestore collection corresponding to the chat
conversation and returns a stream of `QuerySnapshot` objects containing
the latest messages. The messages are ordered by timestamp in descending
order to ensure that the most recent messages are displayed first.

### 3. Message List Widget

The `MessageList` widget is responsible for displaying the chat messages
within the chat view. It utilizes a `StreamBuilder` to listen to changes
in the stream of `QuerySnapshot` objects returned by the
`ChatService.getMessages` method.

Upon receiving new data from the stream, the `StreamBuilder` rebuilds
its child widget, which is a `ListView.builder`. This builder iterates
over the list of messages retrieved from Firestore and constructs
`ChatBubble` widgets to display each individual message.

### 4. Chat Bottom Bar Widget

The `ChatBottomBar` widget provides a user interface for composing and
sending messages. It consists of a text input field for typing messages
and a send button for sending the composed message.

When the send button is pressed, the message text entered by the user is
sent to the `ChatService.sendMessage` method along with the recipient's
user ID. The `ChatService` then adds the message to the Firestore
collection corresponding to the chat conversation.

### 5. Chat Bubble Widget

The `ChatBubble` widget is used to display individual chat messages
within the message list. It formats the message content along with the
sender's name and profile picture in a visually appealing manner.

Each `ChatBubble` widget is constructed with the message data retrieved
from Firestore and displayed accordingly in the chat interface.