Chat Eventual Connectivity, Local Storage and Caching Strategies (#154)
# Eventual Connectivity

Display messages when there's no internet connection. Save messages in a
queue to be sent as soon as internet is established. Notice how the
messages arange themselves in order on both devices.


https://github.com/ISIS3510-202410-Team-13/Flutter/assets/68788933/22372f9e-13b5-4b71-94b6-52cd8cc8bbcc

# Local Storage

Notice that the chats work even offline! It uses Hive for both Chat and
Messages information.


https://github.com/ISIS3510-202410-Team-13/Flutter/assets/68788933/d157bafa-d212-448c-a8e4-660e65b90171

# Caching

The text that the user types in the text field is preserved, even if the
user leaves the app or loses internet connection.


https://github.com/ISIS3510-202410-Team-13/Flutter/assets/68788933/eeb79d71-fcdf-4558-bfb9-7310fa5fba27